Refactored InvoiceService names
